Make functions that can be private private
The `set_bit` and `get_bit` functions do no error checking and are not
supposed to be used directly. The bits should be set/get using a `u64`
where the function guarantees that the index will not be out of bounds.

The function to get this "safe" bit index from a `u64` is also an
internal detail, not used by the consumer.